Move prices to tiers
