+++ src/components/pricing/index.js
@@ -3,5 +3,5 @@ export { default as SidebarLogoContent } from './SidebarLogoContent';
 export { default as SectionHeader } from './SectionHeader';
 export { default as FeatureComparisonTable } from './FeatureComparisonTable';
 export { default as CurrencySwitcher } from './CurrencySwitcher';
-export { TIERS } from './prices';
+export { TIERS } from './tiers';
 export { default as Tier } from './Tier';
